requirements.txt
asgiref==3.4.1
astroid==2.6.6
Django==3.2.6
djangorestframework==3.12.4
isort==5.9.3
jedi==0.18.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
parso==0.8.2
pylint==2.9.6
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
wrapt==1.12.1
